# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Pulse Healthcare API",
        "description": "A hospital booking and management platform connecting patients with healthcare facilities in Ghana. Manages hospital registration, doctor scheduling, appointment booking, and real-time queue tracking.",
        "contact": {
            "name": "Marvinphil Annorbah",
            "url": "https://github.com/psam-717/pulse",
            "email": "mphilannorbah@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "v2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearer-jwt": []
        }
    ],
    "paths": {
...
        "/api/auth/patient/verify-otp": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "verifyOtp",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyOtpRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/auth/patient/signup": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "signup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignupRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/auth/patient/login": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "login_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatientLoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/auth/admin/login": {
            "post": {
                "tags": [
                    "admin-controller"
                ],
                "operationId": "login_2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DoctorLoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/auth/admin/create-doctor": {
            "post": {
                "tags": [
                    "admin-controller"
                ],
                "operationId": "createDoctor",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDoctorRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuthResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "VerifyOtpRequest": {
                "type": "object",
                "properties": {
                    "phone": {
                        "type": "string"
                    },
                    "otp": {
                        "type": "string"
                    }
                }
            },
            "ApiResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "SignupRequest": {
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "ghanaCard": {
                        "type": "string"
                    }
                }
            },
            "PatientLoginRequest": {
                "type": "object",
                "properties": {
                    "identifier": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "DoctorLoginRequest": {
                "type": "object",
                "properties": {
                    "workspaceId": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "CreateDoctorRequest": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "specialization": {
                        "type": "string"
                    },
                    "departmentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "hospitalId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "licenseNumber": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer-jwt": {
                "type": "http",
                "description": "Enter your JWT token. Get one from POST /api/auth/login or /api/auth/register\n\nExample: admin token \u2192 `Bearer eyJhbGciOiJIUzI1NiJ...`",
                "name": "bearer-jwt",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}